refactor : LottoNumbers 클래스 내부 메서드 Lotto로 이관  및 LottoNumbers 클래스 제거
